requirements.txt
Flask
PyWebview